refactor(agent/runtime): extract run lifecycle management into AgentRunLifecycle
`executeAgent.ts` previously owned both flow routing (which flow to run based
on AgentCategory) and run lifecycle management (stream-status transitions,
execution registry tracking, error classification, user notifications, and
resource disposal). The two concerns were entangled in one 612-line file,
making it hard to audit the lifecycle invariants in isolation.

Extract `runFlowWithLifecycle` and `buildTerminalFlowResult` into a dedicated
`AgentRunLifecycle.ts`. `executeAgent.ts` now delegates lifecycle management to
that module and stays focused on orchestration: building the launch context,
routing to the correct flow, and assembling the `AgentFlowResult`.

The exclusive imports (`writeTerminalStatus`, `logSdkError`, `classifyAgentError`,
`INSTRUCTION_ACTION`, `AgentExecutionHandle`, `executionRegistry`,
`EXECUTION_STATUS`, `END_GROUP_STATUS`) move with the code, so `executeAgent.ts`
sheds seven import sites. No behaviour change.
