+++ src/agent/runtime/executeAgent.ts
@@ -1,7 +1,5 @@
 import * as path from 'path';
 
-import { writeTerminalStatus } from '@agent/storage';
-import { logSdkError } from '@agent/trace';
 import type { ToolUseSessionSnapshot } from '@agent/implementations/flows/tooluse/ToolUseSessionTypes';
 import {
   runToolUseFlow,
@@ -22,18 +20,10 @@ import {
 } from '@agent/core/definition/AgentDataclass';
 import { computeDelegationDepthFromStorage } from '@agent/runtime/delegationPolicy';
 import { agentConfigToTaskState } from '@agent/utils/agentConfigToTaskState';
-import {
-  AgentError,
-  classifyAgentError,
-  getSdkErrorMessage,
-} from '@common/errors';
-import { INSTRUCTION_ACTION } from '@eventBus/ProgressEventBus';
+import { AgentError } from '@common/errors';
 import { createChannelTrace } from '@logger';
 import {
   STREAM_STATUS,
-  END_GROUP_STATUS,
-  EXECUTION_STATUS,
-  type EndGroupStatus,
   type StreamTabId,
   type ExecutionId,
   type OutputFileInfo,
@@ -48,8 +38,11 @@ import {
   withExecutionRunContext,
   type AgentLaunchContext,
 } from './AgentLaunchContext';
+import {
+  buildTerminalFlowResult,
+  runFlowWithLifecycle,
+} from './AgentRunLifecycle';
 import { createInterruptCallbacks } from './InterruptManager';
-import { AgentExecutionHandle, executionRegistry } from './executionRegistry';
 import { generateSessionDescription } from './sessionDescription';
 import { getRunStorageService } from './RunStorageService';
 import { StreamStatusService } from './StreamStatusService';
@@ -157,160 +150,6 @@ function createUsageRecordingCallback(
 ): RoundFinalizedCallback {
   return async (run) => {
     await ctx.usageMonitor.recordUsage(run);
-  };
-}
-
-async function runFlowWithLifecycle(
-  ctx: AgentLaunchContext,
-  streamId: StreamTabId,
-  agentName: string,
-  runner: () => Promise<AgentFlowResult>,
-  options?: {
-    isSubagent?: boolean;
-    category?: 'workflow' | 'toolUse';
-    parentStreamId?: StreamTabId;
-    onCompleted?: (result: AgentFlowResult) => void | Promise<void>;
-    onError?: (error: unknown, result: AgentFlowResult) => void | Promise<void>;
-  },
-): Promise<AgentFlowResult> {
-  const category = options?.category ?? 'workflow';
-  const parentStreamId = options?.parentStreamId ?? streamId;
-  const handle = new AgentExecutionHandle(
-    ctx.executionId,
-    parentStreamId,
-    streamId,
-    agentName,
-    category,
-    ctx.runtimeHost,
-  );
-  executionRegistry.track(handle);
-  try {
-    const result = await runner();
-    await options?.onCompleted?.(result);
-    const terminalStatus =
-      result.status === END_GROUP_STATUS.ERROR
-        ? EXECUTION_STATUS.ERROR
-        : EXECUTION_STATUS.COMPLETED;
-    await writeTerminalStatus(ctx.executionId, terminalStatus).catch(() => {});
-
-    executionRegistry.untrack(ctx.executionId);
-    ctx.parentStage.end(result.status);
-
-    if (!StreamStatusService.shouldPreserveOnCompletion(streamId)) {
-      const status =
-        result.status === 'error' ? STREAM_STATUS.ERROR : STREAM_STATUS.STOPPED;
-      StreamStatusService.set(streamId, status, {
-        runtimeHost: ctx.runtimeHost,
-        terminalStatus,
-      });
-    }
-    logger.debug(`Task completed with status: ${result.status}`);
-    return result;
-  } catch (err) {
-    const kind = classifyAgentError(err);
-    const status =
-      kind === 'abort' ? END_GROUP_STATUS.STOPPED : END_GROUP_STATUS.ERROR;
-    const terminalStatus =
-      kind === 'abort' ? EXECUTION_STATUS.INTERRUPTED : EXECUTION_STATUS.ERROR;
-    const streamStatus =
-      kind === 'abort' ? STREAM_STATUS.STOPPED : STREAM_STATUS.ERROR;
-    await writeTerminalStatus(ctx.executionId, terminalStatus).catch(() => {});
-    executionRegistry.untrack(ctx.executionId);
-    const errorMsg = `Error executing agent ${agentName}: ${getSdkErrorMessage(err)}`;
-
-    // Root-agent failures are surfaced in the stream log. Subagent failures
-    // are delivered to the orchestrator below, so avoid adding a second
-    // wrapper error that makes a child failure look like the parent failed.
-    if (kind !== 'abort' && !options?.isSubagent) {
-      logSdkError(ctx.logger, errorMsg, err, {
-        operation: `execute ${agentName}`,
-      });
-    }
-
-    ctx.parentStage.end(status);
-    StreamStatusService.set(streamId, streamStatus, {
-      runtimeHost: ctx.runtimeHost,
-      terminalStatus,
-    });
-
-    // Subagents propagate errors to the orchestrator via FollowUpQueue —
-    // don't show VS Code popups that would confuse the user.
-    if (!options?.isSubagent) {
-      if (kind === 'disk-full') {
-        ctx.runtimeHost.emit('requestShowError', {
-          message: getSdkErrorMessage(err),
-        });
-      } else if (kind === 'missing-api-key') {
-        ctx.runtimeHost.emit('requestShowInstruction', {
-          key: 'missingApiKey',
-          message:
-            'API key not found. Set your API key in the extension settings and run again.',
-          actions: [
-            INSTRUCTION_ACTION.SET_API_KEY,
-            INSTRUCTION_ACTION.OPEN_CONFIGURATION_GUIDE,
-          ],
-          showSuppress: false,
-        });
-      } else if (kind === 'unexpected') {
-        ctx.runtimeHost.emit('requestShowError', {
-          message: errorMsg,
-        });
-      }
-    }
-
-    if (kind === 'abort') {
-      return buildTerminalFlowResult(
-        category,
-        END_GROUP_STATUS.STOPPED,
-        ctx.executionId,
-        streamId,
-      );
-    }
-
-    if (options?.isSubagent) {
-      const result = buildTerminalFlowResult(
-        category,
-        END_GROUP_STATUS.ERROR,
-        ctx.executionId,
-        streamId,
-      );
-      try {
-        await options.onError?.(err, result);
-      } catch (deliveryError) {
-        logger.warn(
-          `Failed to deliver subagent error for ${agentName}: ${getSdkErrorMessage(deliveryError)}`,
-        );
-      }
-      return result;
-    }
-
-    throw new AgentError(errorMsg, { cause: err });
-  } finally {
-    // Release long-lived resources (e.g., WebSocket connections, keepalive intervals)
-    // to prevent leaks when handler instances are discarded after execution.
-    ctx.modelHandler.dispose();
-    // Drop the run-trace subscribers (channel sink + transcript recorder) so
-    // they don't pile up across many agent runs.
-    ctx.disposeTrace();
-  }
-}
-
-function buildTerminalFlowResult(
-  category: 'workflow' | 'toolUse',
-  status: EndGroupStatus,
-  executionId: ExecutionId,
-  streamId: StreamTabId,
-): AgentFlowResult {
-  if (category === 'toolUse') {
-    return { category, status, executionId, streamId };
-  }
-  return {
-    category,
-    status,
-    executionId,
-    streamId,
-    outputs: [],
-    compileFailures: [],
   };
 }
